summit routes implementation + general logic refactors

+++ backend/LogSummitApi.Presentation/Controllers/SummitController.cs
@@ -1,8 +1,11 @@
-﻿using LogSummitApi.Domain.Core.Dto.Summit;
-using LogSummitApi.Domain.Core.Entities;
-using LogSummitApi.Domain.Core.Exceptions.HTTP;
+﻿using LogSummitApi.Application.Core.Extensions;
+using LogSummitApi.Domain.Core.Dto.Summit;
+using LogSummitApi.Domain.Core.Exceptions.Http;
 using LogSummitApi.Domain.Core.Interfaces.Services;
+using LogSummitApi.Presentation.Extensions;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.IdentityModel.Tokens;
 
 namespace LogSummitApi.Presentation.Controllers;
 
@@ -18,32 +21,14 @@ public class SummitController : ControllerBase
     }
 
     [HttpGet("summit")]
-    public async Task<IActionResult> Index(int limit, int offset, string? country)
-    {
-        var summits = await _service.Summit.IndexAsync();
-
-        if (country is not null)
-        {
-            summits = summits.Where(s => s.Country!.Equals(country, StringComparison.CurrentCultureIgnoreCase));
-        }
-        if (offset > 0)
-        {
-            summits = summits.Skip(offset);
-        }
-        if (limit > 0)
-        {
-            summits = summits.Take(limit);
-        }
-
-        return Ok(summits.Select(s => s.ToDto()).ToList()); 
-    }
-
     [HttpGet("summit/user/{userId}")]
-    public async Task<IActionResult> IndexByUser(Guid userId, int limit, int offset, string? country)
+    public async Task<IActionResult> Index(Guid? userId, int limit, int offset, string? country)
     {
         var summits = await _service.Summit.IndexAsync();
 
-        if (userId != Guid.Empty)
+        summits = summits.Where(s => s.IsPublic); // leave out all the private summits
+
+        if (userId != null && userId != Guid.Empty)
         {
             summits = summits.Where(s => s.UserId == userId);
         }
@@ -51,35 +36,32 @@ public class SummitController : ControllerBase
         {
             summits = summits.Where(s => s.Country!.Equals(country, StringComparison.CurrentCultureIgnoreCase));
         }
-        if (offset > 0)
-        {
-            summits = summits.Skip(offset);
-        }
-        if (limit > 0)
-        {
-            summits = summits.Take(limit);
-        }
 
-        return Ok(summits.Select(s => s.ToDto()).ToList()); 
+        return Ok(summits.Paginate(offset, limit));
     }
+
 
     [HttpGet("summit/{summitId}")]
     public async Task<IActionResult> Get(Guid summitId)
     {
         var summit = await _service.Summit.GetAsync(summitId);
 
-        return Ok(summit.ToDto());
+        return Ok(summit);
     }
 
     [HttpGet("summit/valid-countries")]
     public async Task<IActionResult> GetValidCountries()
     {
         return Ok(await _service.Summit.GetValidCountriesAsync());
+
     }
 
-    [HttpPost("summit")]
+    [HttpPost("summit"), Authorize(Policy = "User")]
     public async Task<IActionResult> Create([FromBody] CreateSummitDto createSummitDto)
     {
+        // authenticate the request
+        if (createSummitDto.UserId != this.GetUserIdFromJwt()) throw new NotAuthorized401Exception();
+
         var summit = await _service.Summit.CreateAsync(createSummitDto);
 
         return Created($"/v1/api/summit/{summit.Id}", null);
@@ -90,7 +72,12 @@ public class SummitController : ControllerBase
     {
         try
         {
-            await _service.Summit.UpdateAsync(await _service.Summit.GetAsync(summitId), updateSummitDto);
+            var summit = await _service.Summit.GetAsync(summitId);
+
+            // authenticate the request
+            if (summit.UserId != this.GetUserIdFromJwt()) throw new NotAuthorized401Exception();
+
+            await _service.Summit.UpdateAsync(summit, updateSummitDto);
 
             return NoContent();
         }
@@ -114,6 +101,9 @@ public class SummitController : ControllerBase
     public async Task<IActionResult> Delete(Guid summitId)
     {
         var summit = await _service.Summit.GetAsync(summitId);
+
+        // authenticate the request
+        if (summit.UserId != this.GetUserIdFromJwt()) throw new NotAuthorized401Exception();
 
         await _service.Summit.DeleteAsync(summit);
